requirements.txt
python-dotenv>=1.0.0
rich>=10.0.0
pyinstaller>=5.6.0
